refactor: split validator.py into focused modules to meet 300-line limit (#143)
Extract input validation into input_validator.py and formula parameter
validation into formula_param_validator.py. Reduces validator.py from
334 lines to 226 lines.
